chore: Move EncodedByteBuffer to utils package
This commit moves the `EncodedByteBuffer` class from the `core.encoding` package and places it into the `utils` package.
