+++ common/src/main/java/com/wynntils/utils/EncodedByteBuffer.java
@@ -2,7 +2,7 @@
  * Copyright © Wynntils 2023.
  * This file is released under LGPLv3. See LICENSE for full license details.
  */
-package com.wynntils.core.encoding;
+package com.wynntils.utils;
 
 import com.wynntils.utils.type.UnsignedByte;
 import java.util.ArrayList;
